requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
line-bot-sdk==2.1.0
MarkupSafe==2.1.0
multidict==6.0.2
openpyxl==3.0.9
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
yarl==1.7.2
